requirements.txt
click==6.2
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
ordereddict==1.1
python-dotenv==0.3.0
Werkzeug==0.11.5
wsgiref==0.1.2
gunicorn==19.4.5
q==2.6
